@@ -1,5 +1,6 @@
 import type { Handler } from "express";
 import { allowedTokens } from "./config";
+import { logger } from "../common/logger";
 
 /**
  * Returns a middleware function that checks for a valid token in the request
@@ -16,9 +17,12 @@ export const verifyTokenMiddleware: Handler = (request, response, next) => {
 
   if (matched) return next();
 
-  return response
-    .status(401)
-    .json({
-      error: "Unauthorized: Invalid Authorization. API Token is Required!",
-    });
+  logger.warn(
+    { allowedTokens, current: authHeader },
+    "Unauthorized: Invalid Authorization. API Token is Required!",
+  );
+
+  return response.status(401).json({
+    error: "Unauthorized: Invalid Authorization. API Token is Required!",
+  });
 };
