@@ -1,10 +1,10 @@
 import axios from "axios";
-import { toArray } from "../../common/arrayUtils";
-import type { Fact, FactService, PathCountResults } from "./types";
+// import { toArray } from "../../common/arrayUtils";
+import type { FactAdapter, FactEntity } from "./types";
 
 const FactsConfig = {
-  apiUrl: process.env.API_URL ?? "http://localhost:3000/api/facts",
-  apiToken: process.env.API_TOKEN ?? null,
+  apiUrl: process.env["API_URL"] ?? "http://localhost:3000/api",
+  apiToken: process.env["API_TOKEN"] ?? null,
 };
 
 const client = axios.create({
@@ -19,41 +19,38 @@ const enc = encodeURIComponent;
 /**
  * This is the HTTP client for the FactService.
  */
-const FactApiClient: FactService = {
-  create: (fact) => client.put(`/`, fact),
+const FactApiClient: FactAdapter = {
+  _name: "http",
 
-  updateById: ({ id, ...fact }) =>
-    client.post<Fact[]>(`/${Number(id)}`, fact).then((res) => res.data),
+  set: ({ key, ...fact }) =>
+    client.put(`/facts/${enc(key)}`, fact).then((res) => res.data),
 
-  updateByPathKey: (update, fact) =>
+  get: ({ key }) =>
+    client.get<FactEntity>(`/facts/${enc(key)}`).then((res) => res.data),
+
+  // updateByPathKey: (update, fact) =>
+  //   client
+  //     .post<Fact[]>(`/${enc(update.path)}/${enc(update.key.toString())}`, fact)
+  //     .then((res) => res.data),
+
+  del: ({ key }) => client.delete(`/facts/${enc(key)}`).then((res) => res.data),
+
+  // getPathCounts: () =>
+  //   client.get<PathCountResults>(`/?count=path`).then((res) => res.data),
+
+  find: ({ keyPrefix }) =>
     client
-      .post<Fact[]>(`/${enc(update.path)}/${enc(update.key.toString())}`, fact)
+      .get<FactEntity[]>(`/query/${enc(keyPrefix)}`)
       .then((res) => res.data),
 
-  removeById: (id) => client.delete(`/${id}`).then((res) => res.data),
-
-  getPathCounts: () =>
-    client.get<PathCountResults>(`/?count=path`).then((res) => res.data),
-
-  findFactsByPathKeys: ({ path, key, limit }) =>
-    client
-      .get<Fact[]>(`/`, {
-        params: {
-          path,
-          key: toArray(key),
-          limit: limit ?? 50,
-        },
-      })
-      .then((res) => res.data),
-
-  findAllFactsByPath: ({ path, limit }) =>
-    client
-      .get<Fact[]>(`/${encodeURIComponent(path)}`, {
-        params: {
-          limit: limit ?? 250,
-        },
-      })
-      .then((res) => res.data),
+  // findAllFactsByPath: ({ path, limit }) =>
+  //   client
+  //     .get<Fact[]>(`/${encodeURIComponent(path)}`, {
+  //       params: {
+  //         limit: limit ?? 250,
+  //       },
+  //     })
+  //     .then((res) => res.data),
 };
 
 export default FactApiClient;
